+++ share/ChangeLog.md
@@ -1,5 +1,6 @@
 2024-10-02:
 
+* add support for "liken", "faken", and related verbs
 * add support for "designen"
 
 2024-09-30: